+++ src/main/java/seouldata/seoul_backend/domain/cctv/domain/repository/AnCctvRepository.java
@@ -1,7 +1,7 @@
-package seouldata.seoul_backend.cctv.domain.repository;
+package seouldata.seoul_backend.domain.cctv.domain.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import seouldata.seoul_backend.cctv.domain.entity.AnCctv;
+import seouldata.seoul_backend.domain.cctv.domain.entity.AnCctv;
 
 public interface AnCctvRepository extends JpaRepository<AnCctv, String> {
