+++ src/main/java/seouldata/seoul_backend/domain/cctv/domain/service/AnCctvService.java
@@ -1,16 +1,16 @@
-package seouldata.seoul_backend.cctv.domain.service;
+package seouldata.seoul_backend.domain.cctv.domain.service;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import seouldata.seoul_backend.cctv.application.dto.request.CctvRequest;
-import seouldata.seoul_backend.cctv.application.dto.response.CctvResponse;
-import seouldata.seoul_backend.cctv.domain.entity.AnCctv;
-import seouldata.seoul_backend.cctv.domain.repository.AnCctvRepository;
+import seouldata.seoul_backend.domain.cctv.application.dto.request.CctvRequest;
+import seouldata.seoul_backend.domain.cctv.application.dto.response.CctvResponse;
+import seouldata.seoul_backend.domain.cctv.domain.entity.AnCctv;
+import seouldata.seoul_backend.domain.cctv.domain.repository.AnCctvRepository;
+import seouldata.seoul_backend.common.DistanceUtils;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
@@ -38,7 +38,7 @@ public class AnCctvService {
             double lon = cctv.getLon();
             double lat = cctv.getLat();
 
-            double distance = calculationDistance(userLon, userLat, lon, lat);
+            double distance = DistanceUtils.calculateDistance(userLon, userLat, lon, lat);
 
             if(distance <= 500) {
                 CctvResponse.CctvNearResponse nearCctv = CctvResponse.CctvNearResponse.builder()
@@ -55,21 +55,4 @@ public class AnCctvService {
         return nearCctvs;
 
     }
-
-    // 하버 사인 적용하여 두 점 사이 거리 구하기
-    public double calculationDistance(double userLon, double userLat, double lon, double lat) {
-        double earthRadius = 6371000; // 지구 반지름(미터)
-
-        double dLat = Math.toRadians(lat - userLat);
-        double dLon = Math.toRadians(lon - userLon);
-
-        double a = Math.sin(dLat / 2) * Math.sin(dLat / 2)
-                + Math.cos(Math.toRadians(userLat)) * Math.cos(Math.toRadians(lat))
-                * Math.sin(dLon / 2) * Math.sin(dLon / 2);
-
-        double c = 2 * Math.atan2(Math.sqrt(a), Math.sqrt(1 - a));
-        return earthRadius * c;
-    }
-
-
 }
