+++ src/main/java/seouldata/seoul_backend/domain/cctv/presentation/AnCctvController.java
@@ -1,13 +1,13 @@
-package seouldata.seoul_backend.cctv.presentation;
+package seouldata.seoul_backend.domain.cctv.presentation;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
-import seouldata.seoul_backend.cctv.application.dto.request.CctvRequest;
-import seouldata.seoul_backend.cctv.application.dto.response.CctvResponse;
-import seouldata.seoul_backend.cctv.domain.service.AnCctvService;
+import seouldata.seoul_backend.domain.cctv.application.dto.request.CctvRequest;
+import seouldata.seoul_backend.domain.cctv.application.dto.response.CctvResponse;
+import seouldata.seoul_backend.domain.cctv.domain.service.AnCctvService;
 
 import java.util.List;
